Moved obs table format converter to datastore; integrate the converter in the initialisation of the datastore classes.
